+++ R/nbpmatching-package.R
@@ -62,7 +62,7 @@ NULL
 #'This function should not be called by the user.
 #'
 #'@name nbpMatching-internal
-#'@aliases .requireCachedGenerics initialize,distancematrix-method [<-,distancematrix-method [[<-,distancematrix-method runner runner,data.frame-method full.qom full.qom,data.frame-method
+#'@aliases .requireCachedGenerics initialize,distancematrix-method [<-,distancematrix-method [[<-,distancematrix-method runner runner,data.frame-method full.qom full.qom,data.frame-method show,nonbimatch-method
 #'@exportMethod runner
 #'@exportMethod full.qom
 #'@keywords internal